requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
numpy==1.26.4
pandas==2.2.3
psycopg2-binary==2.9.9
sentence-transformers==2.7.0
tqdm==4.67.1
openai==1.63.0
requests==2.31.0
python-dotenv==1.1.0
neo4j==5.19.0
pinecone-client==3.0.0
